+++ CHANGELOG.md
@@ -8,7 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+- [VSL-25](https://github.com/textnow/vessel/issues/25) Added new `replace` method overload to support passing in the old data model class type
 - [VSL-3](https://github.com/textnow/vessel/issues/3) Added the ability to enable caching for Vessel
+
+### Deprecated
+
+- [VSL-25](https://github.com/textnow/vessel/issues/25) Deprecated `replace` method overload that required the old data model object to be passed in
 
 ## [0.1.1] - 2021-09-10